Refactored and modularised code
