+++ src/config.py
@@ -40,7 +40,7 @@ FACE_DATA_DIR = "/Volumes/T7/vipl_videos/"
 TARGET_SIGNAL_DIR = "/Users/anweshcr7/Downloads/CleanerPPG/VIPL-HR/Cleaned/"
 # SAVE_CSV_PATH = "subject_exclusive_folds.csv"
 SAVE_CSV_PATH = "VIPL_npy.csv"
-ST_MAPS_PATH = "../data/vipl_st_maps/vipl_st_maps_full/"
+ST_MAPS_PATH = "../data/vipl_st_maps/"
 # ST_MAPS_PATH = "/Volumes/Backup Plus/vision/DEAP_emotion/st_maps/"
 CHECKPOINT_PATH = "../checkpoint"
 DATA_PATH = "../data/"
